App.jsx deleted, carousel.jsx compounded
